[DDW-563] Remove unnecessary files
