+++ .prettierignore
@@ -23,4 +23,3 @@ source/renderer/app/i18n/locales/
 source/renderer/app/config/newsfeed-files/
 tests/paper-wallets/e2e/documents/*
 tests/wallets/e2e/documents/*
-utils/restore-keystore/*
