+++ source/main/cardano/utils.js
@@ -6,7 +6,7 @@ import { spawnSync } from 'child_process';
 import { logger } from '../utils/logging';
 import { getTranslation } from '../utils/getTranslation';
 import ensureDirectoryExists from '../utils/ensureDirectoryExists';
-import { decodeKeystore } from '../../../utils/restore-keystore/dist/index';
+import { decodeKeystore } from '../utils/restoreKeystore';
 import type { LauncherConfig } from '../config';
 import type { ExportWalletsMainResponse } from '../../common/ipc/api';
 import type {
@@ -238,6 +238,7 @@ export const exportWallets = async (
   let errors = '';
   try {
     const legacySecretKeyFile = fs.readFileSync(legacySecretKeyPath);
+    // $FlowFixMe
     const rawWallets = await decodeKeystore(legacySecretKeyFile);
     wallets = rawWallets.map((w) => ({
       name: null,
