--- utils/restore-keystore/jest.config.ts
@@ -1,9 +1,0 @@
-module.exports = {
-  "roots": [
-    "<rootDir>",
-  ],
-  "transform": {
-    "^.+\\.spec.ts?$": "ts-jest"
-  },
-  preset: 'ts-jest'
-}
