--- utils/restore-keystore/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "restore-keystore",
-  "version": "1.0.0",
-  "author": "KtorZ <matthias.benkort@iohk.io>",
-  "license": "Apache-2.0",
-  "description": "A utility library for decoding legacy keystore from cardano-sl wallet state directories.",
-  "main": "./dist/index.js",
-  "typings": "./dist/index.d.ts",
-  "dependencies": {
-    "blake2b": "2.1.3",
-    "cbor": "6.0.1"
-  },
-  "devDependencies": {
-    "@types/cbor": "5.0.1",
-    "@types/jest": "26.0.20",
-    "@types/node": "14.14.21",
-    "jest": "26.6.3",
-    "shx": "0.3.2",
-    "ts-jest": "26.4.4",
-    "ts-node": "9.1.1",
-    "typescript": "4.1.3"
-  },
-  "scripts": {
-    "build": "shx rm -rf dist && tsc",
-    "prepare": "npm run build",
-    "test": "jest"
-  }
-}
