--- utils/restore-keystore/tsconfig.eslint.json
@@ -1,8 +1,0 @@
-
-{
-  "extends": "./tsconfig.json",
-  "include": [
-    "src/**/*.ts",
-    "test/**/*.ts"
-  ]
-}
