--- utils/restore-keystore/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
-  "compilerOptions": {
-    "target": "es2016",
-    "module": "commonjs",
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "outDir": "dist",
-    "removeComments": true,
-    "noEmitOnError": true,
-    "strict": true,
-    "noUnusedLocals": false,
-    "esModuleInterop": false,
-    "preserveConstEnums": true,
-    "noImplicitAny": false
-  }
-}
